v8

+++ back-front-base/frontend/src/app/page/home/home.component.ts
@@ -1,10 +1,11 @@
 import { Component, inject, OnInit } from '@angular/core';
-import { BehaviorSubject, catchError, of, switchMap } from 'rxjs';
+import { BehaviorSubject, catchError, map, of, switchMap } from 'rxjs';
 import { AuthService } from '../../service/auth.service';
 import { RequestService } from '../../service/request.service';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { AddRequestModal } from '../../components/add-request-modal/add-request-modal';
-import { User } from '../../entities/user.entity';
+import { User } from '../../service/user.entity';
+import { UserService } from '../../service/user.service';
 
 @Component({
   selector: 'app-home',
@@ -17,33 +18,54 @@ export class HomeComponent implements OnInit {
   protected requestService = inject(RequestService);
   protected authSrv = inject(AuthService);
   private modalService = inject(NgbModal);
+  protected userSrv = inject(UserService);
 
   refresh$ = new BehaviorSubject<void>(undefined);
 
   userRole: string | null = null;
   user: User | null = null;
+  users: User[] = [];
 
-  ngOnInit() {
-    this.user = this.authSrv.getCurrentUser();
-    this.userRole = this.user?.role ?? null;
-  }
+userMap = new Map<string, User>();
 
-  request$ = this.authSrv.isAuthenticated$.pipe(
-    switchMap(isAuth => {
-      if (!isAuth) return of([]);
+ngOnInit() {
+  this.user = this.authSrv.getCurrentUser();
+  this.userRole = this.user?.role ?? null;
 
-      return this.refresh$.pipe(
-        switchMap(() =>
-          this.requestService.list().pipe(
-            catchError(err => {
-              console.error(err);
-              return of([]);
-            })
-          )
+  this.userSrv.list('role1').subscribe(res => {
+    this.users = res;
+    this.userMap = new Map(res.map(u => [u.id!, u]));
+  });
+}
+getUserName(id: string) {
+  console.log('REQUEST ID:', id);
+  console.log('USER MAP KEYS:', Array.from(this.userMap.keys()));
+
+  const u = this.userMap.get(id);
+
+  return u ? `${u.firstName} ${u.lastName}` : 'Utente sconosciuto';
+}
+ request$ = this.authSrv.isAuthenticated$.pipe(
+  switchMap(isAuth => {
+    if (!isAuth) return of([]);
+
+    return this.refresh$.pipe(
+      switchMap(() =>
+        this.requestService.list().pipe(
+          map(list =>
+            list.filter(r =>
+              r.stato !== 'Rifiutato' 
+            )
+          ),
+          catchError(err => {
+            console.error(err);
+            return of([]);
+          })
         )
-      );
-    })
-  );
+      )
+    );
+  })
+);
 
   openAdd() {
     const modalRef = this.modalService.open(AddRequestModal);
@@ -55,18 +77,13 @@ export class HomeComponent implements OnInit {
     }).catch(() => {});
   }
 
-  refreshRequests() {
-  this.requestService.list().subscribe(list => {
-    this.request$ = of(list);
-  });
-}
+ deleteRequest(id: string) {
+  if (!confirm('Vuoi eliminare questa richiesta?')) return;
 
-deleteRequest(id: string) {
   this.requestService.delete(id).subscribe(() => {
-    this.refreshRequests();
+    this.refresh$.next();
   });
 }
-
   approveRequest(id: string) {
     this.requestService.approveRequest(id).subscribe(() => {
       this.refresh$.next();
